:sparkles: Add accessibility resources description on web
